Separate index.js into modules
